requirements.txt
django==2.2.7
djangorestframework
requests
mysqlclient
djoser
